refactor : TwoButtonModal -> ConfirmModal로 이름 변경
