+++ src/components/Home/Header/UserConfig/index.tsx
@@ -3,7 +3,7 @@ import { ARROW } from '@/assets/icons/Arrow';
 import { SYSTEM } from '@/assets/icons/System';
 import { USER } from '@/assets/icons/User';
 import DropDown from '@/components/@common/DropDown';
-import { TwoButtonModal } from '@/components/@common/Modal/TwoButtonModal';
+import { ConfirmModal } from '@/components/@common/Modal/ConfirmModal';
 import UserConfigModal from '@/components/@common/Modal/UserConfigModal';
 import { useEffect, useRef, useState } from 'react';
 import * as Style from './style';
@@ -52,7 +52,7 @@ const UserConfig = () => {
         </Style.UserConfigButton>
       </Style.UserConfig>
       {showLogOutModal && (
-        <TwoButtonModal
+        <ConfirmModal
           title="로그아웃"
           description="로그아웃 하시겠습니까?"
           modalHandler={handleShowLogOutmodal}
